Move view styled components to separate components
