+++ frontend/src/metabase/query_builder/components/view/View/View/index.ts
@@ -1,0 +1,2 @@
+export * from "./View";
+export * from "./View.styled";
