+++ frontend/src/metabase/query_builder/components/view/View/ViewHeaderContainer/index.ts
@@ -1,1 +1,3 @@
 export * from "./ViewHeaderContainer";
+export { QueryBuilderViewHeaderContainer } from "metabase/query_builder/components/view/View/ViewHeaderContainer/ViewHeaderContainer.styled";
+export { BorderedViewTitleHeader } from "metabase/query_builder/components/view/View/ViewHeaderContainer/ViewHeaderContainer.styled";
