+++ frontend/src/metabase/query_builder/components/view/View/ViewMainContainer/index.ts
@@ -1,1 +1,4 @@
 export * from "./ViewMainContainer";
+export { StyledSyncedParametersList } from "metabase/query_builder/components/view/View/ViewMainContainer/ViewMainContainer.styled";
+export { StyledDebouncedFrame } from "metabase/query_builder/components/view/View/ViewMainContainer/ViewMainContainer.styled";
+export { QueryBuilderMain } from "metabase/query_builder/components/view/View/ViewMainContainer/ViewMainContainer.styled";
