+++ frontend/src/metabase/query_builder/components/view/View/ViewNativeQueryEditor/index.ts
@@ -1,1 +1,2 @@
 export * from "./ViewNativeQueryEditor";
+export { NativeQueryEditorContainer } from "metabase/query_builder/components/view/View/ViewNativeQueryEditor/ViewNativeQueryEditor.styled";
